rename: 관리자페이지 카드뉴스 폴더 생성
