+++ src/pages/index.js
@@ -50,8 +50,8 @@ export { Admin } from '../components/layout/Admin';
 export { AdminLogin } from './admin/AdminLogin';
 export { BoardManagement } from './admin/board/BoardManagement';
 export { CreateBoard } from './admin/board/CreateBoard';
-export { CardnewsList } from './admin/CardnewsList';
-export { CreateCardnews } from './admin/CreateCardnews';
+export { CardnewsList } from './admin/cardnews/CardnewsList';
+export { CreateCardnews } from './admin/cardnews/CreateCardnews';
 export { Dashboard } from './admin/Dashboard';
 export { ReportManagement } from './admin/ReportManagement';
 export { SignupManagement } from './admin/SignupManagement';
